+++ src/models/items/basic/rune.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Rune {
     pub isrune: bool,
-    pub tier: i32,
+    pub tier: f32,
     #[serde(rename = "type")]
     pub rune_type: String,
 }
